+++ src/oscar/Utils/Perf.h
@@ -16,21 +16,21 @@ namespace osc
     // internal details needed for `OSC_PERF` to work
     namespace detail
     {
-        int64_t AllocateMeasurementID(
+        size_t AllocateMeasurementID(
             std::string_view label,
             std::string_view filename,
             unsigned int line
         );
 
         void SubmitMeasurement(
-            int64_t id,
+            size_t id,
             PerfClock::time_point start,
             PerfClock::time_point end
         );
 
         class PerfTimer final {
         public:
-            explicit PerfTimer(int64_t id) :
+            explicit PerfTimer(size_t id) :
                 m_ID{id}
             {
             }
@@ -43,7 +43,7 @@ namespace osc
                 SubmitMeasurement(m_ID, m_Start, PerfClock::now());
             }
         private:
-            int64_t m_ID;
+            size_t m_ID;
             PerfClock::time_point m_Start = PerfClock::now();
         };
     }
@@ -52,5 +52,5 @@ namespace osc
 #define OSC_PERF_TOKENPASTE(x, y) x##y
 #define OSC_PERF_TOKENPASTE2(x, y) OSC_PERF_TOKENPASTE(x, y)
 #define OSC_PERF(label) \
-    static int64_t const OSC_PERF_TOKENPASTE2(s_TimerID, __LINE__) = osc::detail::AllocateMeasurementID(label, osc::ExtractFilename(__FILE__), __LINE__); \
+    static size_t const OSC_PERF_TOKENPASTE2(s_TimerID, __LINE__) = osc::detail::AllocateMeasurementID(label, osc::ExtractFilename(__FILE__), __LINE__); \
     osc::detail::PerfTimer const OSC_PERF_TOKENPASTE2(timer, __LINE__) (OSC_PERF_TOKENPASTE2(s_TimerID, __LINE__));
